fix: bugs around property matcher handling
Before, we were doing all kinds of broken things:

1. Undercloning. I.e., we were mutating deep objects in the user’s data
when a jsonpath matcher applied to them (because we only did a shallow
clone).

2. Overcloning. If the user didn’t have property matchers, we still did
an unecessary shallow clone on the data being snapshotted. This only
happened if it was an object without custom serializers, so it was
probably harmless, but still uneccessary.

3. We skipped all property matchers when a custom serializer was in
use, when we really only needed to log a warning.